+++ mdbx_load.c
@@ -17,7 +17,7 @@
 /// \copyright SPDX-License-Identifier: Apache-2.0
 /// \author Леонид Юрьев aka Leonid Yuriev <leo@yuriev.ru> \date 2015-2026
 
-#define MDBX_BUILD_SOURCERY 183c1d490880cf26022164bc32c0ead32de47d91b204a36f92cb2acb1ce1bda1_v0_13_12_154_g9ca22c60
+#define MDBX_BUILD_SOURCERY 66b709f51a75e4768d4900bbe763b83ba1e0cc0c2de0a37691c626d926a05779_v0_13_12_164_g7aca4951
 
 #define LIBMDBX_INTERNALS
 #define MDBX_DEPRECATED
@@ -3830,6 +3830,7 @@ int main(int argc, char *argv[]) {
   MDBX_cursor *mc = nullptr;
   MDBX_dbi dbi;
   char *envname = nullptr;
+  char *input_file = nullptr;
   int envflags = MDBX_SAFE_NOSYNC | MDBX_ACCEDE, putflags = MDBX_UPSERT;
   bool rescue = false;
   bool purge = false;
@@ -3865,11 +3866,7 @@ int main(int argc, char *argv[]) {
       putflags |= MDBX_APPEND;
       break;
     case 'f':
-      if (freopen(optarg, "r", stdin) == nullptr) {
-        if (!quiet)
-          fprintf(stderr, "%s: %s: open: %s\n", prog, optarg, mdbx_strerror(errno));
-        exit(EXIT_FAILURE);
-      }
+      input_file = optarg;
       break;
     case 'n':
       envflags |= MDBX_NOSUBDIR;
@@ -3897,8 +3894,20 @@ int main(int argc, char *argv[]) {
     }
   }
 
-  if (optind != argc - 1)
-    usage();
+  if (optind != argc - 1) {
+    /* The man page promises the standard input by default and requires
+     * dbpath as the last argument. When '-f' consumed the only remaining
+     * argument as its file, treat that argument as the dbpath and fall
+     * back to stdin (issue #50). */
+    if (input_file && optind == argc) {
+      envname = input_file;
+      input_file = nullptr;
+    } else {
+      usage();
+    }
+  } else {
+    envname = argv[optind];
+  }
 
 #if defined(_WIN32) || defined(_WIN64)
   SetConsoleCtrlHandler(ConsoleBreakHandlerRoutine, true);
@@ -3913,7 +3922,13 @@ int main(int argc, char *argv[]) {
   signal(SIGTERM, signal_handler);
 #endif /* !WINDOWS */
 
-  envname = argv[optind];
+  if (input_file) {
+    if (freopen(input_file, "r", stdin) == nullptr) {
+      if (!quiet)
+        fprintf(stderr, "%s: %s: open: %s\n", prog, input_file, mdbx_strerror(errno));
+      exit(EXIT_FAILURE);
+    }
+  }
   if (!quiet) {
     printf("mdbx_load %s (%s, T-%s)\nRunning for %s...\n", mdbx_version.git.describe, mdbx_version.git.datetime,
            mdbx_version.git.tree, envname);